Refactor(graduation.*) : 서비스 모듈 통일성 맞추기
