+++ src/main/java/wiseSWlife/wiseSWlife/service/graduation/exam/Exam.java
@@ -1,11 +1,11 @@
-package wiseSWlife.wiseSWlife.service.graduation.scraping;
+package wiseSWlife.wiseSWlife.service.graduation.exam;
 
 import wiseSWlife.wiseSWlife.dto.graduation.ExamTable;
 import wiseSWlife.wiseSWlife.dto.graduation.form.ExamForm;
 
 import java.io.IOException;
 
-public interface ExamScraping {
+public interface Exam {
 
     ExamTable scraping(String intCookie) throws IOException, InterruptedException;
