+++ src/main/java/wiseSWlife/wiseSWlife/service/graduation/totalAcceptanceStatus/TotalAcceptanceStatus.java
@@ -1,10 +1,10 @@
-package wiseSWlife.wiseSWlife.service.graduation.scraping;
+package wiseSWlife.wiseSWlife.service.graduation.totalAcceptanceStatus;
 
 import wiseSWlife.wiseSWlife.dto.graduation.TotalAcceptanceStatusTable;
 
 import java.io.IOException;
 
-public interface TotalAcceptanceStatusScraping {
+public interface TotalAcceptanceStatus {
 
     TotalAcceptanceStatusTable scrapping(String intCookie) throws IOException, InterruptedException;
 }
