+++ src/test/java/wiseSWlife/wiseSWlife/service/graduation/scrapingImpl/ExamTest.java
@@ -1,14 +1,5 @@
 package wiseSWlife.wiseSWlife.service.graduation.scrapingImpl;
 
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import wiseSWlife.wiseSWlife.dto.member.Member;
-import wiseSWlife.wiseSWlife.service.graduation.scraping.Exam;
-import wiseSWlife.wiseSWlife.service.login.SimpleLoginService;
-
-import java.io.IOException;
-import java.util.Map;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 //public class ExamTest {
